{"schema":"decodex/commit/1","summary":"Split state marker and attempt modules","authority":"manual"}
