+++ apps/decodex/src/state/run_activity_marker/storage.rs
@@ -1,6 +1,6 @@
 use std::{
 	fs::{self, OpenOptions},
-	io::{ErrorKind, Write},
+	io::{ErrorKind, Write as _},
 	path::Path,
 	process,
 	sync::atomic::{AtomicU64, Ordering},
@@ -8,10 +8,11 @@ use std::{
 
 use crate::{
 	prelude::{Result, eyre},
-	state::{CodexAccountActivitySummary, RUN_ACTIVITY_MARKER_FILE},
+	state::{
+		CodexAccountActivitySummary, RUN_ACTIVITY_MARKER_FILE,
+		run_activity_marker::record::RunActivityMarkerRecord,
+	},
 };
-
-use super::RunActivityMarkerRecord;
 
 static RUN_ACTIVITY_MARKER_WRITE_SEQUENCE: AtomicU64 = AtomicU64::new(0);
 
@@ -38,12 +39,15 @@ pub(crate) fn read_run_activity_marker_record(
 			"process_id" => marker.process_id = value.parse::<u32>().ok(),
 			"host_boot_id" => marker.host_boot_id = Some(value.to_owned()),
 			"process_start_identity" => marker.process_start_identity = Some(value.to_owned()),
-			"last_activity_unix_epoch" =>
-				marker.last_activity_unix_epoch = value.parse::<i64>().ok(),
-			"last_protocol_activity_unix_epoch" =>
-				marker.last_protocol_activity_unix_epoch = value.parse::<i64>().ok(),
-			"last_progress_unix_epoch" =>
-				marker.last_progress_unix_epoch = value.parse::<i64>().ok(),
+			"last_activity_unix_epoch" => {
+				marker.last_activity_unix_epoch = value.parse::<i64>().ok()
+			},
+			"last_protocol_activity_unix_epoch" => {
+				marker.last_protocol_activity_unix_epoch = value.parse::<i64>().ok()
+			},
+			"last_progress_unix_epoch" => {
+				marker.last_progress_unix_epoch = value.parse::<i64>().ok()
+			},
 			"current_operation" => marker.current_operation = Some(value.to_owned()),
 			"thread_id" => marker.thread_id = Some(value.to_owned()),
 			"turn_id" => marker.turn_id = Some(value.to_owned()),
@@ -54,24 +58,30 @@ pub(crate) fn read_run_activity_marker_record(
 			"effective_model" => marker.effective_model = Some(value.to_owned()),
 			"effective_model_provider" => marker.effective_model_provider = Some(value.to_owned()),
 			"effective_cwd" => marker.effective_cwd = Some(value.to_owned()),
-			"effective_approval_policy" =>
-				marker.effective_approval_policy = Some(value.to_owned()),
-			"effective_approvals_reviewer" =>
-				marker.effective_approvals_reviewer = Some(value.to_owned()),
+			"effective_approval_policy" => {
+				marker.effective_approval_policy = Some(value.to_owned())
+			},
+			"effective_approvals_reviewer" => {
+				marker.effective_approvals_reviewer = Some(value.to_owned())
+			},
 			"effective_sandbox_mode" => marker.effective_sandbox_mode = Some(value.to_owned()),
-			"child_agent_activity" =>
-				marker.child_agent_activity = serde_json::from_str(value).ok(),
+			"child_agent_activity" => {
+				marker.child_agent_activity = serde_json::from_str(value).ok()
+			},
 			"protocol_activity" => marker.protocol_activity = serde_json::from_str(value).ok(),
 			"account" => marker.account = serde_json::from_str(value).ok(),
-			"accounts" =>
+			"accounts" => {
 				if let Ok(accounts) = serde_json::from_str(value) {
 					marker.accounts = accounts;
-				},
-			"retry_budget_attempt_count" =>
-				marker.retry_budget_attempt_count = value.parse::<i64>().ok(),
+				}
+			},
+			"retry_budget_attempt_count" => {
+				marker.retry_budget_attempt_count = value.parse::<i64>().ok()
+			},
 			"retry_kind" => marker.retry_kind = Some(value.to_owned()),
-			"retry_ready_at_unix_epoch" =>
-				marker.retry_ready_at_unix_epoch = value.parse::<i64>().ok(),
+			"retry_ready_at_unix_epoch" => {
+				marker.retry_ready_at_unix_epoch = value.parse::<i64>().ok()
+			},
 			_ => {},
 		}
 	}
@@ -111,9 +121,10 @@ fn preserve_current_run_account_marker_fields(
 		return;
 	};
 	let keep_current_account = match next.account.as_ref() {
-		Some(next_account) =>
+		Some(next_account) => {
 			account_marker_observed_unix_epoch(&current_account)
-				> account_marker_observed_unix_epoch(next_account),
+				> account_marker_observed_unix_epoch(next_account)
+		},
 		None => true,
 	};
